Automated rollback of commit ea2e7970c87c452acc70b44e8f3a0191302f07a4.
*** Reason for rollback ***

Causes spurious "undefined name" errors; see b/167383811

*** Original change description ***

bazel syntax: fix bug in comprehension resolution

Following Python3, the operand of the first for clause
is treated specially: it is resolved outside the comprehension
block.

The scope of all for variables is the entire comprehension block,
so the operand of a later for clause may refer to its own variable.
It is possible to use this dynamically for a loop-carried dependence.

Thanks to @stepancheg for identifying the problem.

PiperOrigin-RevId: 329507606
Bazel-Commit=23a677e61427753dbbb17f18730bb8dcad57558d
Upstream-Source=bazelbuild/bazel
GitOrigin-RevId: 23a677e61427753dbbb17f18730bb8dcad57558d