Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java bazel syntax: fix bug in comprehension resolution
(Second attempt at commit ea2e7970c87c452acc70b44e8f3a0191302f07a4, rolled back in commit 23a677e61427753dbbb17f18730bb8dcad57558d.)

Following Python3, the operand of the first for clause
is treated specially: it is resolved outside the comprehension
block.

The scope of all for variables is the entire comprehension block,
so the operand of a later for clause may refer to its own variable.
It is possible to use this dynamically for a loop-carried dependence.

Thanks to @stepancheg for identifying the problem.

RELNOTES:
 The Starlark interpreter now correctly emits an error
 if the operand of the first loop in a list comprehension
 refers to a variable bound by a later loop, such as y in
 this example:
   [e1 for x in f(y) in e2 for y in e3] # error: undefined y
                  ^
 This may cause latent dynamic errors to become static errors.
PiperOrigin-RevId: 329566001
GitOrigin-RevId: b48b7064c4104a70626832aeb024e17a22a13ac7