starlark annot: delete StarlarkGlobalLibrary (step 1)
Step 1: remove uses in Bazel, and all its semantics.
It is still used by Copybara.

Docgen relied upon it to indicate which classes should
be scanned for StarlarkMethod-annotated methods to be
included in the "globals" doc. This role is now served
by the new docgen.annot.DocumentMethods annotation.
(I suspect it could be eliminated with smarter heuristics,
but scanning all classes for such methods causes docgen
to crash.)

Any class may provide annotated methods without implementing
StarlarkValue. You can pass it to addMethods to treat its
methods as Starlark standalone functions, but the class
instance is not a legal Starlark value.

A StarlarkValue class may have a StarlarkBuiltin annotation
to provide its type name and documentation, but this is optional.
Without it, the type name is based on that of the Java class.

Also, improve an assertion in docgen.

PiperOrigin-RevId: 332488054
Bazel-Commit=89c1398fabea704e0c31a61a86210af81fc52c0f
Upstream-Source=bazelbuild/bazel
GitOrigin-RevId: 89c1398fabea704e0c31a61a86210af81fc52c0f