Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java starlark: improve expectation hygiene in ScriptTest
Prior to this change, ScriptTest inherited a fuzzy syntax for
expectations from its Python-based predecessor: the content
of a ### comment was interpreted either as a substring, or
as a regular expression.

This change causes it to interpret the content of the comment
as a regular expression, always. Some expectations were updated.

PiperOrigin-RevId: 338258986
GitOrigin-RevId: 07f092e7133e811dcda9d6eac11115ee1fbbee6b