Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java starlark: clean up Dict construction
Dict.copyOf used to call fromJava for each map value.
This was a hack to support Bazel's attr.string_list_dict type
(a Map<String, List<String>>) so that lists in dicts would be
recursively converted; but it has no place here.
Now, fromJava makes only a shallow copy, and Dict.copyOf
asserts checkValid on all keys and values.

Dict.Builder is a builder of Dicts, mutable or immutable.
This avoids a number of hash table copies.

Various places in the loading phase assume that Starlark.fromJava
is the appropriate conversion from the internal form of rule
attribute values to Starlark form. That is now incorrect.
Attribute.valueToStarlark is the correct conversion, and those
places now call it instead.

PiperOrigin-RevId: 340350758
GitOrigin-RevId: b74f1601586d4437409a28d7a4d2e907c8625c79