starlark: remove StarlarkList.Builder
I added it in commit 0c56a0636cf594bed2351d4ca5fe8db330cb531f in pursuit of (a foolish) consistency with Dict,
but it has an aliasing bug, and the fix requires it to make a copy,
which defeats the already marginal benefit of copy avoidance except
in the case when the result is immutable, which is not an important
use-case outside of tests.

Also, ensure that toArray() returns an Object[], not a subclass,
so that clients can insert arbitrary values into it without ArrayStoreException.

PiperOrigin-RevId: 342106898
Bazel-Commit=f9eb69882feaa55094fd0844126ebfd34c3f59c5
Upstream-Source=bazelbuild/bazel
GitOrigin-RevId: f9eb69882feaa55094fd0844126ebfd34c3f59c5