Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java starlark tests: use assert_fails(lambda) instead of '###' annotations
The new form avoids the need to put all errors at the end of a chunk.
Instead, they can be grouped with the tests of successful operations.

Not every instance was replaced; the ### notation is more convenient
when exercising statements, or in situations where the regexp needs
a lot of escaping. Also, a new --- chunk can help isolate side effects.

No behavioral changes.

PiperOrigin-RevId: 347390509
GitOrigin-RevId: ddf95df15d06d3dd3ab291e0cc2b75e8499a07b0