Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java starlark: int: ignore base prefix unless matches explicit base
Previously, when int was called with an explicit base,
it would report an error if the digit string starts
with a base prefix for a different base, such as int("0b101", 16).
Now, it uses the base prefix only if it matches the requested
base, so the example above would return 0x0b101, as would
int("0x0b101", 16).

See github.com/google/starlark-go/pull/344

PiperOrigin-RevId: 355479386
GitOrigin-RevId: 8fc77874b07a94d8fc4223bf7de20254838420b1