Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Fix type of `arguments` param in `ctx.actions.run()` and `run_shell()`
In both cases, we want to restrict the type to a Sequence in Java (to avoid an
unsafe cast) and to document it as a sequence of unspecified type in Starlark
API docs (unfortunately, ParamType annotations currently don't allow us to
express a Sequence containing 2 different allowed types possibly intermixed).

This prevents a Bazel crash when trying to use a scalar string value for the
`arguments` param in `ctx.actions.run_shell()`.

RELNOTES: None.
PiperOrigin-RevId: 368918401
GitOrigin-RevId: 8a43e0cb973ab979da8140bc31fd94651537ce5b