Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Add Starlark mechanism for default_copts and defaults_hdrs_check
We add a new native computed default class which can only be seen from builtins. This ensures that not only are we able to get the value of default_copts and default_hdrs_check as set in the package() rule, but that when the values in package() changes, the dependency server is able to compute the affected targets properly.

This is not a long term mechanism and should be replaced by something more generic as described in b/200065655#comment3

This CL also adds two additional methods to CcStarlarkInternal, init_make_variables and create_linkstamp.

RELNOTES:none
PiperOrigin-RevId: 415258162
GitOrigin-RevId: ee38b630272128376b4491d794ee28ca0e871f9d