Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Have `Dict#contents` switch to an `ImmutableMap` when the `Dict` is frozen.
This implements the idea I documented in a TODO in https://github.com/bazelbuild/bazel/commit/24ec910935a7f6c2da1f22db48e198ced0711eaf. The approach is to introduce an optional mechanism for Starlark values to get notified when their `Mutability` gets frozen, and to use that mechanism in `Dict`. In implementing this idea, I noticed yet another idea for saving memory. I added a TODO for that, and I also added a TODO for a code unification noticed by arostovtsev@ during code review.

For a `bazel build --nobuild //very/large:target` invocation with many retained `Dict` instances, this approach reduces retained heap by ~1%. There is a ~0.65% increase in CPU usage but no measurable increase in wall time, so I think this tradeoff is very much so worth it.

PiperOrigin-RevId: 433563032
GitOrigin-RevId: 5576979669383c2380c11b59beaa2f8263ff1f56