Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Introduce new Starlark.positionalOnlyCall() function for calls with only positional params.
Use it in the many suitable places where Starlark.call() is used so far.
Next steps will be to introduce something corresponding in StarlarkCallable and implement a more efficient implementation thereof in StarlarkFunction and possibly BuiltinFunction.
Further steps could be to allow optimized access paths from Eval.evalCall() to StarlarkFunction and, again possibly, to BuiltinFunction, that directly populate the locals array or the argument vector for the native call.

PiperOrigin-RevId: 703075286
Change-Id: I8fbda39645884a32831958b1fa6b9a1603b9567f
GitOrigin-RevId: fdfdbd8a669ddef1866a350e53b89966a363c09d