Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Add `--incompatible_enforce_starlark_utf8`
If enabled (or set to `error`), fail if Starlark files are not UTF-8 encoded. If set to `warning` (the default), emits a warning instead.

Bazel already assumes that Starlark files are UTF-8 encoded for e.g. filenames in actions executed remotely. This flag doesn't affect this, it only makes encoding failures more visible.

Work towards #374

Closes #24944.

PiperOrigin-RevId: 721513249
Change-Id: I1d3363168c6cd5d37abf96e0401e34866b6679d7
GitOrigin-RevId: e7934ce92a8f74e0395fb8086582c0781945196c