Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Fix Unicode encoding issues in Bazel's use of Starlark
Bazel internally uses `String` as a container for raw bytes assumed to be UTF-8, which differs from ordinary usage of `String` as a container for UTF-16 characters. This requires special implementations of certain Starlark functions that care about the notion of a "character":

* `{l,r,}strip` must not strip non-ASCII whitespace as it may be part of a UTF-8-encoded non-whitespace character.
* `json.decode` has to emit UTF-8 bytes rather than UTF-16 characters.

To avoid affecting other users of the Starlark interpreter, a new `StarlarkSemantics.INTERNAL_BAZEL_ONLY_UTF_8_BYTE_STRINGS` setting, defaulting to false but overridden to true for Bazel, determines which of the two behaviors to adopt.

Compatibility is verified by running all script-based tests under both values of the setting.

Closes #24417.

PiperOrigin-RevId: 733329982
Change-Id: I3e1605be28a844ab52a3239a2f753a29d4eb217a
GitOrigin-RevId: fcd3d1976cdace885c43fb5b00da3e98150eb77a