Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Clarify invariant for Binding.first
Originally, `Resolver.Binding.first` (an `Identifier`) was set to null for predeclared/universal bindings, which have no binding occurrence in the syntax tree. However, as of unknown commit this was no longer true, and these bindings have `first` set to a non-binding usage.

This CL removes the ambiguity by fixing `first` to not be marked `@Nullable` and by adding a boolean to distinguish these two cases. Note that we can't simply return to setting this field null for predeclareds/universals, because it is also how the symbol knows its own name.

I tried to add some test cases to demonstrate how the old contract led to incorrect error messages in `toplevelRebinding()` and `globalLocalConflict()`, but I don't think it's actually possible to exercise those code paths, so I settled for just fixing the code without adding tests.

Broken off from discussion in unknown commit.

Work toward #27371.

PiperOrigin-RevId: 824802160
Change-Id: Iabe3efcc5745f554fbecea2cc6677eec06003545
GitOrigin-RevId: ec7be9f9c66a1df0c49a2bfee5e73360148ec669