Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Let `NodeVisitor` subclasses customize whether to visit non-symbol identifiers
Previously, if a subclass wanted to skip over field names and keyword arguments, it had to filter them out explicitly by defining the corresponding `visit()` overloads. This CL makes this a customizable behavior of the base class.

In the future it can also filter out the names of symbols `load()`ed from other modules, since these are not symbols of the current module. (Right now we don't traverse them regardless.)

Work toward #27728.

PiperOrigin-RevId: 841433515
Change-Id: I8e6c024d1751d0e5728ff25e79bc3a8dae17c9cd
GitOrigin-RevId: 5f520bc181934c066b575db797070a6765fe48fb