Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Process type-related syntax in resolver
When the new `FileOptions#resolveTypeSyntax` is enabled, the resolver now creates bindings for symbols introduced by `TypeAliasStatement`, and it sets bindings on `Identifier`s appearing within type expressions. Disabling this flag (the current default) allows type expressions to be tolerated and ignored, despite the lack of definitions for these types in the environment.

Generic type variables (`S` and `T` in `def f[S, T](...)` and `type Foo[S, T] = ...`) still remain unbound and unsupported.

Updated resolver tests, including fixing failures due to the mock environment only supporting a single predeclared symbol `pre`.

Refactoring: `Resolver#assign()` is meant to be used in the second pass to set bindings on all identifiers appearing in a read context within the LHS of an assignment (e.g. `d` and `k` in `d[k] = v`), while skipping over identifiers in write contexts (which should've had their bindings set by the first pass). It's a no-op when called on a simple identifier. Yet we were doing this anyway in `VarStatement`, simply to mark that we didn't forget to operate on that child. I created a couple assertion helpers that are more readable as an alternative to calling `assign()`, and also made `assign()` and some other `visit()` methods use those helpers.

Removed an unused behavior on the return value of `EvaluationTestCase#newModuleHook`.

Work toward #27728.

PiperOrigin-RevId: 842822348
Change-Id: I3ea2a43b73f0663ec539ef69af3f4260b2511705
GitOrigin-RevId: b64c5119068cb55980369fddc3d65614928f387c