Synced-From: bazelbuild/bazel/src/main/java/net/starlark/java Add `TypeChecker.java`, a static type checking visitor
This takes the type annotation information saved in the `Resolver.Binding`s of an AST's `Identifier`s, and validates that the AST uses these types correctly.

The name matches an existing dynamic typechecking file in the `eval/` package, but we can disambiguate that in a future CL. A little stub logic is added to `StarlarkType.java`; in the future we should migrate all the abstract supertype/subtype logic to there.

`Any` is exposed to the user as a valid type annotation identifier.

Work toward #28037.

PiperOrigin-RevId: 846962357
Change-Id: If2abf273c4424814646b82528cd913057e382a85
GitOrigin-RevId: 7d4ae5da75f9aa58885d096c0ba0752deb43f27b